% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{crop_voronois}
\alias{crop_voronois}
\title{Découper les polygones de Voronoi pour respecter les limites d'un polygone}
\usage{
crop_voronois(voronoi_shp, outline)
}
\arguments{
\item{voronoi_shp}{polygones de Voronoi (objet \code{SpatialPolygons*})}

\item{outline}{contour à respecter (objet \code{SpatialPolygons*})}
}
\value{
un objet \code{SpatialPolygons}
}
\description{
Découper les polygones de Voronoi pour respecter les limites d'un polygone
}

