% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{géocode des adresses grâce à l'API BANO}
\usage{
geocode(data, adresses, codeINSEE = NULL)
}
\arguments{
\item{data}{un \code{dataframe} contenant, a minima, les adresses à géocoder}

\item{adresses}{nom de la colonne contenant les adresses à géocoder}

\item{codeINSEE}{nom d'une colonne contenant le code INSEE de la commune. Si \code{NULL} (par défaut), ce paramètre n'est pas utilisé.}
}
\value{
un dataframe avec des colonnes supplémentaires avec la géolocalisation
}
\description{
géocode des adresses grâce à l'API BANO
}

