% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{voronoipolygons}
\alias{voronoipolygons}
\title{Create Voronoi polygons}
\usage{
voronoipolygons(x)
}
\arguments{
\item{x}{un objet de type \code{Spatial*}, ou bien un dataframe dont les deux première colonnes comportent les coordonnées x et y des points à partir desquels générer les polygones de Voronoi.}
}
\description{
Create Voronoi polygons
}

