% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{create_shapes}
\alias{create_shapes}
\title{Crée des polygones de Voronoi à partir de coordonnées géographiques lon/lat}
\usage{
create_shapes(df, ID, x = "longitude", y = "latitude", threshold = NA,
  threshold_score = "result_score", clip_by_chull = FALSE)
}
\arguments{
\item{df}{un dataframe avec les coordonnées géographiques des points ainsi qu'un identifiant par polygone à constituer}

\item{ID}{nom de la colonne contenant les identifiants}

\item{x}{nom de la colonne contenant les longitudes}

\item{y}{nom de la colonne contenant les latitudes}

\item{threshold}{seuil de confiance dans les données géolocalisées à respecter. Si \code{NA} (défaut), pas de seuil.}

\item{threshold_score}{nom de la colonne contenant le degré de confiance dans les données géolocalisées. Par défaut, \code{"result_score"}, nom donné par l'API adresses}

\item{clip_by_chull}{valeur logique. Si \code{TRUE}, les polygones de Voronoi sont réduits à leur enveloppe complexe (le pavage ne couvre donc pas l'intégralité de l'espace mais se limite aux espaces sur lesquels il y a des adresses). Cela est utile par exemple lorsque d'importants espaces non habités existent. \code{FALSE} par défaut.}
}
\value{
un objet \code{SpatialPolygonsDataFrame} contenant les polygones de Voronoi recherchés
}
\description{
Crée des polygones de Voronoi à partir de coordonnées géographiques lon/lat
}

